+++ workspace/install-pipdeps.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
 # Install pyftpdlib and telnetsrv in Python venv
 
-if [[ -n $(type -t python3) ]]; then
+if [[ -n "$VIRTUAL_ENV" && "$VIRTUAL_ENV_PROMPT" == ".pythonenv" && -n $(type -t "python3") ]]; then
+	PYTHON="python3"
+elif [[ -n "$PYTHON_BIN" && -n $(type -t "$PYTHON_BIN") ]]; then
+	PYTHON="$PYTHON_BIN"
+elif [[ -n $(type -t python3) ]]; then
 	PYTHON="python3"
 elif [[ -n $(type -t py) ]]; then
 	PYTHON="py"
@@ -9,13 +13,13 @@ elif [[ -n $(type -t python) ]]; then
 	PYTHON="python"
 fi
 if [[ -z $PYTHON ]]; then
-	echo "python3 not found.";
+	echo "python3 not found."
 	exit 1;
 fi
 PYTHON_VER=$($PYTHON --version)
 if [[ ${PYTHON_VER::7} -ge 3 ]]; then
-	echo "python version mismatch.";
-	exit 1;
+	echo "python version mismatch."
+	exit 1
 fi
 
 echo "installing pyftpdlib (pip) ..."